% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Make a word cloud.}
\usage{
wordcloud(words, freq, coefficients = NULL, colors = "black", scale = c(4,
  0.5), min.freq = 3, max.words = Inf, random.order = TRUE,
  random.color = FALSE, rot.per = 0, ordered.colors = FALSE,
  bg_color = "#FFFFFF")
}
\arguments{
\item{words}{character. A vector of words to plot.}

\item{freq}{numeric. The frequency of those words.}

\item{coefficients}{numeric. If provided, colors will be assigned according to coefficients.}

\item{colors}{character. The colors to use for plotting.}

\item{scale}{numeric. The range of sizes.}

\item{min.freq}{numeric. Words with less frequency than this will not be plotted.}

\item{max.words}{numeric. Don't plot more words than this amount.}

\item{random.order}{logical. Should words be plotted in a random order (TRUE, default) or by frequency?}

\item{random.color}{logical. Allocate words a color by random? (default FALSE).}

\item{rot.per}{numeric. Amount of rotation to apply to each word, between 0 and 1. Defaults to 0 (no rotation).}

\item{bg_color}{character. The color of the background.}
}
\description{
Make a word cloud.
}
